samples: bpf: Add simple sample with CORE
Example checks if received metadata is generic, based on
the last 12 bytes of data_meta.

Example also showcases new CO-RE functionality and is able to
acquire BTF id and type id for both vmlinux and module types.

On the user side, xdp program is detached before termination
is executed, before that trace_pipe is printed for every packet.

Signed-off-by: Michal Swiatkowski <michal.swiatkowski@intel.com>
Signed-off-by: Larysa Zaremba <larysa.zaremba@intel.com>