bpf: samples: Patch module BTF id into bpf insns
Implementation of btf__obj_id() would probably need to change to a
separate bpf syscall case, because just before commiting I've noticed
that this implementation has problems with vmlinux BTF :(.

Furthermore, syscall implementation would be useful when we'll need to
access ids data from a userspace program that's not a libbpf loader
(for example, when using AF_XDP).

Everything aside this function should be OK and can be reviewed.

Signed-off-by: Larysa Zaremba <larysa.zaremba@intel.com>