libbpf: Acquire vmlinux BTF ID from libbpf
Previous version of patching-in BTF ID did not handle vmlinux BTF case,
because unlike kernel module BTFs it was loaded into libbpf from
an actual file, not from an anonymous file descriptor.

Unfortunately, file descriptor is the only thing that links libbpf btf
to a source kernel btf and therefore to BTF obj id.

For vmlinux BTF, there existed no way to know its BTF ID even at the
moment of creation, therefore the main part of this commit is
implementation of such functionality. I've decided to implement
getting a btf_info, because getting ID application is too limited to
create a separate syscall case. The only problem I see now with this is
that currently it reuses syscall bpf_attr struct from another syscall
case.
Refactoring this would require a lot of coding, so I'd rather not do it
until someone confirms it's neccessary.

Also, I am still not sure about how do we determine if the BTF
is vmlinux.
For  now it just assumes that every BTF without the base_btf is vmlinux,
but that could also be some local or override BTF.
For now I see 2 possible solutions:
- Just return '0' if btf is not a kernel BTF, I suppose vmlinux is
  supposed to be the only kernel BTF without base
- Add 'btf_id' field to libbpf btf struct, initialize it at the moment
  of creation for both vmlinux and module BTFs