libbpf: Use kernel BTF id in userspace
Introduce id field to the userspace btf struct,
it allows to clearly identify btf's in-kernel counterpart.

Unify the way vmlinux and module BTFs are loaded
by encapsulating part of module BTF loading code
in a function that picks the first BTF after an id,
which complies with criteria.

For vmlinux, try to acquire it in the new way,
resort to old fs way, if not successful.

Signed-off-by: Larysa Zaremba <larysa.zaremba@intel.com>