libbpf: Patch module BTF id into bpf insns
Return both type id and BTF id from bpf_core_type_id_kernel().
Earlier only type id was returned despite the fact that llvm
has enabled the 64 return type for this instruction [1].
This was done as a preparation to the patch [2], which
also strongly served as a inspiration for this implementation.

Signed-off-by: Larysa Zaremba <larysa.zaremba@intel.com>