samples: bpf: pass a struct to sample_install_xdp()
In order to be able to pass more flags and/or other options to
sample_install_xdp() from userland programs built ontop of this
framework, make it consume a const pointer to a structure with
all the parameters needed to initialize the sample instead of a
set of parameters which doesn't scale.
Adjust all the samples accordingly.

Signed-off-by: Alexander Lobakin <alexandr.lobakin@intel.com>